minivt_pipeline/requirements.txt
python-dotenv
pydantic>=2.7
requests
pydub
